CAMEL-19732 - Support Pipe in CRDs, replacing KameletBinding
keeping KameletBinding working to let transition times for users
