--- dsl/camel-jbang/camel-jbang-core/src/main/resources/templates/binding-uri-kamelet.yaml.tmpl
@@ -1,16 +1,0 @@
-apiVersion: camel.apache.org/v1alpha1
-kind: KameletBinding
-metadata:
-  name: {{ .Name }}
-spec:
-  source:
-    uri: {{ .Source }}
-    #properties:
-      #key: "value"
-{{ .Steps }}
-  sink:
-    ref:
-      kind: Kamelet
-      apiVersion: camel.apache.org/v1
-      name: {{ .Sink }}
-    {{ .SinkProperties }}
