+++ dsl/camel-jbang/camel-jbang-core/src/main/resources/templates/pipe-kamelet-uri.yaml.tmpl
@@ -1,0 +1,16 @@
+apiVersion: camel.apache.org/v1
+kind: Pipe
+metadata:
+  name: {{ .Name }}
+spec:
+  source:
+    ref:
+      kind: Kamelet
+      apiVersion: camel.apache.org/v1
+      name: {{ .Source }}
+    {{ .SourceProperties }}
+{{ .Steps }}
+  sink:
+    uri: {{ .Sink }}
+    #properties:
+      #key: "value"
