+++ dsl/camel-jbang/camel-jbang-core/src/main/resources/templates/pipe-uri-kamelet.yaml.tmpl
@@ -1,14 +1,12 @@
-apiVersion: camel.apache.org/v1alpha1
-kind: KameletBinding
+apiVersion: camel.apache.org/v1
+kind: Pipe
 metadata:
   name: {{ .Name }}
 spec:
   source:
-    ref:
-      kind: Kamelet
-      apiVersion: camel.apache.org/v1
-      name: {{ .Source }}
-    {{ .SourceProperties }}
+    uri: {{ .Source }}
+    #properties:
+      #key: "value"
 {{ .Steps }}
   sink:
     ref:
